[#2] refactor: 패키지 구조 변경
- adapter를 최상단으로.
- usecase 인터페이스 추가.
