+++ src/main/java/com/baedal/customer/adapter/presentation/controller/CustomerController.java
@@ -1,10 +1,10 @@
-package com.baedal.customer.presentation.controller;
+package com.baedal.customer.adapter.presentation.controller;
 
-import com.baedal.customer.application.service.CustomerAuthenticationService;
-import com.baedal.customer.application.service.CustomerSignupService;
-import com.baedal.customer.presentation.request.LoginRequest;
-import com.baedal.customer.presentation.request.SignupRequest;
-import com.baedal.customer.presentation.response.LoginResponse;
+import com.baedal.customer.adapter.presentation.request.LoginRequest;
+import com.baedal.customer.adapter.presentation.request.SignupRequest;
+import com.baedal.customer.adapter.presentation.response.LoginResponse;
+import com.baedal.customer.application.port.in.CustomerAuthenticationUsecase;
+import com.baedal.customer.application.port.in.CustomerSignupUsecase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -17,20 +17,20 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class CustomerController {
 
-  private final CustomerAuthenticationService customerAuthenticationService;
+  private final CustomerAuthenticationUsecase customerAuthenticationUsecase;
 
-  private final CustomerSignupService customerSignupService;
+  private final CustomerSignupUsecase customerSignupUsecase;
 
   @PostMapping("/v0/login")
   public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest request) {
-    String token = customerAuthenticationService.authenticate(request.email(), request.password());
+    String token = customerAuthenticationUsecase.authenticate(request.email(), request.password());
     LoginResponse response = new LoginResponse(token);
     return ResponseEntity.ok(response);
   }
 
   @PostMapping("/v0/signup")
   public ResponseEntity<Void> singup(@RequestBody SignupRequest request) {
-    customerSignupService.signup(request.email(), request.nickname(), request.password());
+    customerSignupUsecase.signup(request.email(), request.nickname(), request.password());
     return ResponseEntity.noContent().build();
   }
 }
