+++ src/main/java/com/baedal/customer/adapter/presentation/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.baedal.customer.presentation.request;
+package com.baedal.customer.adapter.presentation.request;
 
 public record LoginRequest(String email, String password) {
