+++ src/main/java/com/baedal/customer/adapter/presentation/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.baedal.customer.presentation.request;
+package com.baedal.customer.adapter.presentation.request;
 
 public record SignupRequest(String email, String nickname, String password) {
