+++ src/main/java/com/baedal/customer/adapter/presentation/response/LoginResponse.java
@@ -1,0 +1,5 @@
+package com.baedal.customer.adapter.presentation.response;
+
+public record LoginResponse(String token) {
+
+}
