+++ src/main/java/com/baedal/customer/application/port/in/CustomerAuthenticationUsecase.java
@@ -1,0 +1,6 @@
+package com.baedal.customer.application.port.in;
+
+public interface CustomerAuthenticationUsecase {
+
+  String authenticate(String email, String password);
+}
