+++ src/main/java/com/baedal/customer/application/port/in/CustomerSignupUsecase.java
@@ -1,0 +1,6 @@
+package com.baedal.customer.application.port.in;
+
+public interface CustomerSignupUsecase {
+
+  void signup(String email, String nickname, String rawPassword);
+}
