+++ src/main/java/com/baedal/customer/application/service/CustomerAuthenticationService.java
@@ -1,6 +1,7 @@
 package com.baedal.customer.application.service;
 
-import com.baedal.customer.application.port_out.TokenServicePort;
+import com.baedal.customer.application.port.in.CustomerAuthenticationUsecase;
+import com.baedal.customer.application.port.out.TokenServicePort;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class CustomerAuthenticationService {
+public class CustomerAuthenticationService implements CustomerAuthenticationUsecase {
 
   private final PasswordEncoder passwordEncoder;
   private final UserDetailsService userDetailsService;
