--- src/main/java/com/baedal/customer/presentation/response/LoginResponse.java
@@ -1,5 +1,0 @@
-package com.baedal.customer.presentation.response;
-
-public record LoginResponse(String token) {
-
-}
